Fix various issues raised by CI and bump minimum MW requirement to 1.39
