+++ .github/workflows/ci.yml
@@ -7,22 +7,30 @@ on:
 jobs:
   test:
     name: "PHPUnit: MW ${{ matrix.mw }}, PHP ${{ matrix.php }}"
+    continue-on-error: ${{ matrix.experimental }}
 
     strategy:
+      fail-fast: false
       matrix:
         include:
-          - mw: 'REL1_37'
-            php: '8.0'
-          - mw: 'REL1_38'
-            php: '8.0'
-          - mw: 'REL1_38'
-            php: '8.1'
           - mw: 'REL1_39'
-            php: '8.1'
-          - mw: 'REL1_39'
-            php: '8.2'
+            php: 8.1
+            experimental: false
           - mw: 'REL1_40'
-            php: '8.2'
+            php: 8.1
+            experimental: false
+          - mw: 'REL1_41'
+            php: 8.2
+            experimental: false
+          - mw: 'REL1_42'
+            php: 8.2
+            experimental: false
+          - mw: 'REL1_43'
+            php: 8.3
+            experimental: false
+          - mw: 'master'
+            php: 8.4
+            experimental: true
 
     runs-on: ubuntu-latest
 
@@ -46,7 +54,7 @@ jobs:
             mediawiki
             !mediawiki/extensions/
             !mediawiki/vendor/
-          key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v5
+          key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v6
 
       - name: Cache Composer cache
         uses: actions/cache@v4
@@ -91,15 +99,13 @@ jobs:
 
 
   PHPStan:
-    name: "PHPStan: MW ${{ matrix.mw }}, PHP ${{ matrix.php }}"
+    name: "Static Analysis: MW ${{ matrix.mw }}, PHP ${{ matrix.php }}"
 
     strategy:
       matrix:
         include:
-          - mw: 'REL1_37'
-            php: '8.0'
-          - mw: 'REL1_38'
-            php: '8.1'
+          - mw: 'REL1_43'
+            php: '8.3'
 
     runs-on: ubuntu-latest
 
@@ -123,7 +129,7 @@ jobs:
             mediawiki
             mediawiki/extensions/
             mediawiki/vendor/
-          key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v5
+          key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v6
 
       - name: Cache Composer cache
         uses: actions/cache@v4
@@ -155,73 +161,8 @@ jobs:
       - name: PHPStan
         run: cd extensions/WikibaseExport && php vendor/bin/phpstan analyse --error-format=checkstyle --no-progress | cs2pr
 
-
-
-
-  Psalm:
-      name: "Psalm: MW ${{ matrix.mw }}, PHP ${{ matrix.php }}"
-
-      strategy:
-        matrix:
-          include:
-            - mw: 'REL1_37'
-              php: '8.0'
-            - mw: 'REL1_38'
-              php: '8.1'
-
-      runs-on: ubuntu-latest
-
-      defaults:
-          run:
-              working-directory: mediawiki
-
-      steps:
-          - name: Setup PHP
-            uses: shivammathur/setup-php@v2
-            with:
-                php-version: ${{ matrix.php }}
-                extensions: mbstring
-                tools: composer, cs2pr
-
-          - name: Cache MediaWiki
-            id: cache-mediawiki
-            uses: actions/cache@v4
-            with:
-                path: |
-                    mediawiki
-                    mediawiki/extensions/
-                    mediawiki/vendor/
-                key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v5
-
-          - name: Cache Composer cache
-            uses: actions/cache@v4
-            with:
-                path: ~/.composer/cache
-                key: composer_static_analysis
-
-          - uses: actions/checkout@v4
-            with:
-                path: EarlyCopy
-
-          - name: Install MediaWiki
-            if: steps.cache-mediawiki.outputs.cache-hit != 'true'
-            working-directory: ~
-            run: bash EarlyCopy/.github/workflows/installMediaWiki.sh ${{ matrix.mw }} WikibaseExport
-
-          - uses: actions/checkout@v4
-            with:
-                path: mediawiki/extensions/WikibaseExport
-
-          - name: Composer allow-plugins
-            run: composer config --no-plugins allow-plugins.composer/installers true
-
-          - run: composer update
-
-          - name: Composer install
-            run: cd extensions/WikibaseExport && composer install --no-progress --no-interaction --prefer-dist --optimize-autoloader
-
-          - name: Psalm
-            run: cd extensions/WikibaseExport && pwd && php vendor/bin/psalm --config=psalm.xml --shepherd --stats
+      - run: php vendor/bin/psalm --shepherd --stats
+        if: true
 
 
 
@@ -232,8 +173,8 @@ jobs:
     strategy:
       matrix:
         include:
-          - mw: 'REL1_39'
-            php: '8.2'
+          - mw: 'REL1_43'
+            php: '8.3'
 
     runs-on: ubuntu-latest
 
@@ -263,12 +204,16 @@ jobs:
         uses: actions/cache@v4
         with:
           path: ~/.composer/cache
-          key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v5
+          key: mw_${{ matrix.mw }}-php${{ matrix.php }}_v6
+
+      - uses: actions/checkout@v4
+        with:
+          path: EarlyCopy
 
       - name: Install MediaWiki
         if: steps.cache-mediawiki.outputs.cache-hit != 'true'
         working-directory: ~
-        run: curl https://gist.githubusercontent.com/JeroenDeDauw/49a3858653ff4b5be7ec849019ede06c/raw/installMediaWiki.sh | bash -s ${{ matrix.mw }} WikibaseExport
+        run: bash EarlyCopy/.github/workflows/installMediaWiki.sh ${{ matrix.mw }} WikibaseExport
 
       - uses: actions/checkout@v4
         with:
@@ -288,6 +233,6 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-node@v4
         with:
-          node-version: 16
+          node-version: 22
       - run: npm install
       - run: npm run test
