+++ tests/EntryPoints/SearchEntitiesApiTest.php
@@ -128,6 +128,9 @@ class SearchEntitiesApiTest extends WikibaseExportIntegrationTest {
 	}
 
 	public function testNoMatchesReturnsEmptyResult(): void {
+		$this->skipOnPhp81AndLater();
+		$this->saveProperty( self::INSTANCE_OF_ID, 'string', 'instance of' );
+
 		$response = $this->executeHandler(
 			WikibaseExportExtension::searchEntitiesApiFactory(),
 			new RequestData( [
